feat(commands): moved cors and migrate commands to click;
- Moved `fix-cors` and `migrate` sub-command callbacks to util script.
- Implemented `fix-cors` and `migrate` sub-commands with click.
